Use native Convex Swift SDK (ConvexMobile) in the mobile app
Replace the hand-rolled HTTP client with the official convex-swift SDK:
- Add the convex-swift package (ConvexMobile) to Package.swift.
- New ConvexService wraps ConvexClientWithAuth<String> (live WebSocket, auth,
  reconnection); exposes query (first value of a subscription), reactive
  subscribe, and mutation/action.
- Token-based AuthProvider + TokenStore; AuthManager now drives the SDK's
  login()/logout()/loginFromCache() (Clerk/dev token feeds the store).
- MirrorAPI rewritten to use [String: ConvexEncodable?] args; numeric args sent
  as Double to match v.number(); usage counts decoded as Double.
- friendlyMessage() unwraps ClientError.ConvexError payloads for the UI.
- JSONValue slimmed to a Codable helper (data export + error parsing).
- Docs (README, PLATFORMS.md) updated; tests updated.

Verified against the convex-swift source: ConvexClientWithAuth<T>: ConvexClient,
subscribe/mutation/action signatures, and ClientError cases. Backend unchanged.

+++ MirrorApp/Sources/MirrorModel/Models.swift
@@ -285,10 +285,11 @@ public struct ConversationThread: Codable, Sendable {
     public var messages: [MirrorMessage]
 }
 
-/// Response of `settings:getAiUsageEstimate`.
+/// Response of `settings:getAiUsageEstimate`. Counts are modelled as `Double`
+/// because Convex `v.number()` is float64; format them as integers in the UI.
 public struct AiUsageEstimate: Codable, Sendable {
-    public var calls: Int
-    public var inputTokens: Int
-    public var outputTokens: Int
+    public var calls: Double
+    public var inputTokens: Double
+    public var outputTokens: Double
     public var estimatedCostUsd: Double
 }
